requirements.txt
streamlit>=1.28.0
httpx>=0.26.0
python-dotenv>=1.0.0
pandas>=2.0.0
